+++ chap_core/database/model_templates_and_config_tables.py
